Install GHC from the bindist
We install GHC using the binary release directly because installation of
GHC via ghcup or via stack in the formula context is not reliable.

Specifially the GHC installation with ghcup or stack
fails *sometimes* (with about 1/5 frequency) when trying to create a
temporary directory within the temporary HOME directory used by the
formula build.

The GHC installer works reliably when invoked directly from the bindist